CanvasImage系モジュールをファイル分割した
