+++ src/js/game-object/battery-selector/view/battery-merter.ts
@@ -4,7 +4,7 @@ import * as THREE from "three";
 import { HorizontalAnimationMesh } from "../../../mesh/horizontal-animation";
 import { SimpleImageMesh } from "../../../mesh/simple-image-mesh";
 import { Resources } from "../../../resource";
-import { CANVAS_IMAGE_IDS } from "../../../resource/canvas-image";
+import { CANVAS_IMAGE_IDS } from "../../../resource/canvas-image/ids";
 import { findTextureOrThrow } from "../../../resource/find-texture-or-throw";
 import { TEXTURE_IDS } from "../../../resource/texture/ids";
 import { BatterySelectorModel } from "../model";
