+++ src/js/game-object/battery-selector/view/battery-plus.ts
@@ -3,7 +3,7 @@ import * as THREE from "three";
 
 import { SimpleImageMesh } from "../../../mesh/simple-image-mesh";
 import type { ResourcesContainer } from "../../../resource";
-import { CANVAS_IMAGE_IDS } from "../../../resource/canvas-image";
+import { CANVAS_IMAGE_IDS } from "../../../resource/canvas-image/ids";
 import type { GameObjectAction } from "../../action/game-object-action";
 import type { PushDetector } from "../../push-detector";
 import { circlePushDetector } from "../../push-detector/circle-push-detector";
