+++ src/js/game-object/gauge/view/battery-gauge-unit.ts
@@ -2,7 +2,7 @@ import * as THREE from "three";
 
 import { SimpleImageMesh } from "../../../mesh/simple-image-mesh";
 import { Resources } from "../../../resource";
-import { CANVAS_IMAGE_IDS } from "../../../resource/canvas-image";
+import { CANVAS_IMAGE_IDS } from "../../../resource/canvas-image/ids";
 import { BatteryGaugeUnitProps } from "../model/battery-gauge-unit-model";
 
 /** バッテリーユニットゲージ幅（ピクセル） */
