+++ src/js/resource/canvas-image/ids.ts
@@ -1,0 +1,25 @@
+/** キャンバス用画像IDリストをあつめたもの */
+export const CANVAS_IMAGE_IDS = {
+  BATTERY_BUTTON: "BATTERY_BUTTON",
+  BATTERY_PLUS: "BATTERY_PLUS",
+  BATTERY_MINUS: "BATTERY_MINUS",
+  BATTERY_METER: "BATTERY_METER",
+  BATTERY_NEEDLE: "BATTERY_NEEDLE",
+  BATTERY_LABEL_ATTACK: "BATTERY_LABEL_ATTACK",
+  BATTERY_LABEL_DEFENSE: "BATTERY_LABEL_DEFENSE",
+  TURN_INDICATOR: "TURN_INDICATOR",
+  BURST_BUTTON: "BURST_BUTTON",
+  BURST_BUTTON_LABEL: "BURST_BUTTON_LABEL",
+  PILOT_BUTTON: "PILOT_BUTTON",
+  PILOT_BUTTON_LABEL: "PILOT_BUTTON_LABEL",
+  BIG_BUTTON_DISABLED: "BIG_BUTTON_DISABLED",
+  SMALL_BUTTON_DISABLED: "SMALL_BUTTON_DISABLED",
+  HP_BAR: "HP_BAR",
+  HP_BAR_BACK: "HP_BAR_BACK",
+  BATTERY_GAUGE: "BATTERY_GAUGE",
+  BATTERY_GAUGE_BACK: "BATTERY_GAUGE_BACK",
+  TIME_SCALE_BUTTON: "TIME_SCALE_BUTTON",
+  TIME_SCALE_100: "TIME_SCALE_100",
+  TIME_SCALE_050: "TIME_SCALE_050",
+  TIME_SCALE_025: "TIME_SCALE_025",
+};
